Refactored App.js: separated Firebase config into another file firebaseConfig.js to leep things organized.
